failure support on handler

+++ src/main/java/com/github/ricardocomar/camunda/mockservice/model/Failure.java
@@ -15,6 +15,6 @@ public class Failure {
 
     private Integer retryTimes;
 
-    private Integer retryTimeout;
+    private Long retryTimeout;
 
 }
